Re-added styles back to the component rather than in the styles folder.
