+++ components/gallery/Photo.js
@@ -1,6 +1,7 @@
 import React from "react";
-import { Image, TouchableOpacity } from "react-native";
-import styles from "../../styles/styles";
+import { StyleSheet, Image, TouchableOpacity, Dimensions } from "react-native";
+
+const window = Dimensions.get("window");
 
 const Photo = ({ photo }) => (
   <TouchableOpacity style={styles.photoThumbContainer}>
@@ -8,4 +9,17 @@ const Photo = ({ photo }) => (
   </TouchableOpacity>
 );
 
+const styles = StyleSheet.create({
+  photoThumbContainer: {
+    width: window.width / 3,
+    justifyContent: "center",
+    alignItems: "center",
+    height: window.width / 3
+  },
+  thumbnail: {
+    marginTop: 5,
+    width: window.width / 3.1,
+    height: window.width / 3.1
+  }
+});
 export default Photo;
